Test for 'Bareword "__CLASS__"' parsing bug (#147)
See https://github.com/stevan/p5-mop-redux/issues/147

It's possible that this is not a bug, but if so, it's very mysterious and
needs a better error message.

To replicate this bug, eval a class into existence either by eval'ing it or
'eval "use $class"'; (using a class directly is fine). The class must contain
a __CLASS__ token and you must NOT be in package main when you eval.